refactor: 컨트롤러 테스트 하나로 통합
