+++ src/test/java/cart/controller/ProductControllerTest.java
@@ -1,69 +1,174 @@
 package cart.controller;
 
+import cart.dao.ProductDao;
 import cart.dto.ProductRequest;
-import cart.service.ProductService;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import io.restassured.RestAssured;
+import io.restassured.http.ContentType;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 
-import java.util.List;
-
-import static cart.fixture.ProductFixture.FIRST_PRODUCT;
-import static cart.fixture.ProductFixture.SECOND_PRODUCT;
-import static org.mockito.BDDMockito.given;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.hamcrest.Matchers.is;
 
 @SuppressWarnings("NonAsciiCharacters")
-@WebMvcTest(ProductController.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 class ProductControllerTest {
-    @MockBean
-    private ProductService productService;
+    @LocalServerPort
+    private int port;
 
     @Autowired
-    private MockMvc mockMvc;
+    NamedParameterJdbcTemplate namedParameterJdbcTemplate;
 
     @Autowired
-    private ObjectMapper objectMapper;
+    ProductDao productDao;
 
-    @Test
-    void 상품을_조회한다() throws Exception {
-        given(productService.findAll()).willReturn(List.of(FIRST_PRODUCT.RESPONSE, SECOND_PRODUCT.RESPONSE));
-
-        mockMvc.perform(get("/products"))
-                .andExpect(status().isOk())
-                .andDo(print())
-                .andReturn();
+    @BeforeEach
+    void setUp() {
+        productDao.deleteAll();
+        namedParameterJdbcTemplate.getJdbcTemplate().execute("ALTER TABLE product ALTER COLUMN id RESTART WITH 1");
+        RestAssured.port = port;
     }
 
     @Test
-    void 상품을_생성한다() throws Exception {
-        mockMvc.perform(post("/products")
-                        .content(objectMapper.writeValueAsString(FIRST_PRODUCT.REQUEST))
-                        .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isCreated())
-                .andDo(print());
+    void 상품_목록을_조회한다() {
+        RestAssured.given()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when()
+                .get("/products")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value())
+                .extract();
     }
 
     @Test
-    void 상품을_수정한다() throws Exception {
-        final ProductRequest productRequest = new ProductRequest("홍고", "aaaa", 10000);
-
-        mockMvc.perform(patch("/products/1")
-                        .content(objectMapper.writeValueAsString(productRequest))
-                        .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isNoContent())
-                .andDo(print());
+    void 상품을_생성한다() {
+        final ProductRequest productRequest = new ProductRequest("아벨", "aaaa", 10000);
+        RestAssured.given().log().all()
+                .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                .body(productRequest)
+                .when().post("/products")
+                .then().log().all()
+                .statusCode(HttpStatus.CREATED.value())
+                .header("Location", "/products/1");
     }
 
     @Test
-    void 상품을_삭제한다() throws Exception {
-        mockMvc.perform(delete("/products/1"))
-                .andExpect(status().isNoContent());
+    void 상품을_수정한다() {
+        final ProductRequest productRequest = new ProductRequest("아벨", "aaaa", 10000);
+        RestAssured.given().log().all()
+                .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                .body(productRequest)
+                .when().patch("/products/1")
+                .then().log().all()
+                .statusCode(HttpStatus.NO_CONTENT.value());
+    }
+
+    @Test
+    void 상품을_삭제한다() {
+        RestAssured.given().log().all()
+                .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                .when().delete("/products/1")
+                .then().log().all()
+                .statusCode(HttpStatus.NO_CONTENT.value());
+    }
+
+    @ParameterizedTest(name = "{displayName} : name = {0}")
+    @NullAndEmptySource
+    void 상품_이름이_null_또는_empty일_시_예외_발생(final String name) {
+        final ProductRequest productRequest = new ProductRequest(name, "asd", 10);
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(productRequest)
+                .when().post("/products")
+                .then().log().all()
+                .statusCode(HttpStatus.BAD_REQUEST.value())
+                .contentType(ContentType.JSON)
+                .body("message", is("[ERROR] 상품 이름을 입력해주세요."));
+    }
+
+    @Test
+    void 상품_이름_길이가_255초과일때_예외_발생() {
+        final ProductRequest productRequest = new ProductRequest("a".repeat(256), "asd", 10);
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(productRequest)
+                .when().post("/products")
+                .then().log().all()
+                .statusCode(HttpStatus.BAD_REQUEST.value())
+                .contentType(ContentType.JSON)
+                .body("message", is("[ERROR] 상품 이름은 255자까지 입력가능합니다."));
+    }
+
+    @ParameterizedTest(name = "{displayName} : name = {0}")
+    @NullAndEmptySource
+    void 이미지_URL이_null_또는_empty일_시_예외_발생(final String imageUrl) {
+        final ProductRequest productRequest = new ProductRequest("홍고", imageUrl, 10);
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(productRequest)
+                .when().post("/products")
+                .then().log().all()
+                .statusCode(HttpStatus.BAD_REQUEST.value())
+                .contentType(ContentType.JSON)
+                .body("message", is("[ERROR] 이미지 URL을 입력해주세요."));
+    }
+
+    @Test
+    void 이미지_URL_길이가_255초과일때_예외_발생() {
+        final ProductRequest productRequest = new ProductRequest("아벨", "a".repeat(256), 10);
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(productRequest)
+                .when().post("/products")
+                .then().log().all()
+                .statusCode(HttpStatus.BAD_REQUEST.value())
+                .contentType(ContentType.JSON)
+                .body("message", is("[ERROR] 이미지 URL은 255자까지 입력가능합니다."));
+    }
+
+    @Test
+    void 가격이_null일_시_예외_발생() {
+        final ProductRequest productRequest = new ProductRequest("홍고", "홍고", null);
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(productRequest)
+                .when().post("/products")
+                .then().log().all()
+                .statusCode(HttpStatus.BAD_REQUEST.value())
+                .contentType(ContentType.JSON)
+                .body("message", is("[ERROR] 가격을 입력해주세요."));
+    }
+
+    @Test
+    void 가격이_1원_미만일때_예외_발생() {
+        final ProductRequest productRequest = new ProductRequest("아벨", "a", 0);
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(productRequest)
+                .when().post("/products")
+                .then().log().all()
+                .statusCode(HttpStatus.BAD_REQUEST.value())
+                .contentType(ContentType.JSON)
+                .body("message", is("[ERROR] 가격의 최소 금액은 1원입니다."));
+    }
+
+    @Test
+    void 가격이_천만원_초과일때_예외_발생() {
+        final ProductRequest productRequest = new ProductRequest("아벨", "a", 10_000_001);
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(productRequest)
+                .when().post("/products")
+                .then().log().all()
+                .statusCode(HttpStatus.BAD_REQUEST.value())
+                .contentType(ContentType.JSON)
+                .body("message", is("[ERROR] 가격의 최대 금액은 1000만원입니다."));
     }
 }
