+++ src/test/java/cart/fixture/ProductFixture.java
@@ -6,16 +6,16 @@ import cart.dto.ProductResponse;
 
 public class ProductFixture {
     public static final class FIRST_PRODUCT {
-        public static final Product PRODUCT = new Product("홍고", "https://ca.slack-edge.com/TFELTJB7V-U04M4NFB5TN-e18b78fabe81-512", 10_000_000);
-        public static final Product PRODUCT_WITH_ID = new Product(1L, "홍고", "https://ca.slack-edge.com/TFELTJB7V-U04M4NFB5TN-e18b78fabe81-512", 10_000_000);
-        public static final ProductRequest REQUEST = new ProductRequest("홍고", "https://ca.slack-edge.com/TFELTJB7V-U04M4NFB5TN-e18b78fabe81-512", 10_000_000);
+        public static final Product PRODUCT = new Product("hongo", "url1", 10000);
+        public static final Product PRODUCT_WITH_ID = new Product(1L, "hongo", "url1", 10000);
+        public static final ProductRequest REQUEST = new ProductRequest("hongo", "url1", 10000);
         public static final ProductResponse RESPONSE = ProductResponse.from(PRODUCT_WITH_ID);
     }
 
     public static final class SECOND_PRODUCT {
-        public static final Product PRODUCT = new Product("아벨", "https://ca.slack-edge.com/TFELTJB7V-U04LMNLQ78X-a7ef923d5391-512", 10_000_000);
-        public static final Product PRODUCT_WITH_ID = new Product(2L, "아벨", "https://ca.slack-edge.com/TFELTJB7V-U04LMNLQ78X-a7ef923d5391-512", 10_000_000);
-        public static final ProductRequest REQUEST = new ProductRequest("아벨", "https://ca.slack-edge.com/TFELTJB7V-U04LMNLQ78X-a7ef923d5391-512", 10_000_000);
+        public static final Product PRODUCT = new Product("abel", "url2", 10_000_000);
+        public static final Product PRODUCT_WITH_ID = new Product(2L, "abel", "url2", 10_000_000);
+        public static final ProductRequest REQUEST = new ProductRequest("abel", "url2", 10_000_000);
         public static final ProductResponse RESPONSE = ProductResponse.from(PRODUCT_WITH_ID);
     }
 }
